requirements.txt
psycopg2-binary
python-dotenv
loguru